Fix realignment error on the boundary
When checking to see if a base at position [i] is preceded by an
underscore, we check [i - 1]. But, if the base is at position 0, this
checks position -1 (i.e. the last element). Whoops!

This adds a check when the base is right on the boundary of the region
of interest to avoid wrapping around.