Refactor presenter.py methods to use Alignments
Reference and read strings were being passed around in pairs by many of
the methods in presenter.py. The Alignment value object represents that
pair, so we can refactor those methods to use it, instead.

Specifically:

* _get_cas9_region_in_reference takes an Alignment
* _get_new_cas9_regions takes and returns Alignments
* _realign takes an returns Alignments
* get_representations returns an Alignment
* The _realign_cas9_region method can be removed altogether

We're also adding a few type annotation docstrings to the methods to
make things a bit more explicit.