Implement a CSV uploader
We'd like to be able to provide a downloadable CSV of results to our
users. This commit defines a class that handles CSV uploading.

We'd like our filenames to be hard to guess, which should make them
kinda-sorta harder to scrape. I just tacked a timestamp onto the ends of
the uploaded files, including milliseconds, which should do the trick.